Move bake functions to export
